[fix] limit the number of objects returned to 1000 by default
The `each` method would otherwise iterate over all files in the bucket.
This can potentially (and in practice did) result in an enormous list of 100k+ files
which would take an inordinate amount of time to load and use a lot of memory.
Fix this issue by limiting the number of returned objects to 1000 (by default).